As it turns out the 'Heroku Scheduler' is not guaranteed to run on time. Or even run at all in rare situations.
A delay in executing 'Heroku Scheduler' may cause the worker process to outlive its intended schedule.

This adds a way for the worker process to check the schedule itself. This way the Scheduler is mostly used to
return life to the process but not to force it to halt.

Since this can make debugging harder if you chose to test locally just when the process is supposed to die,
 I am also adding a command line argument to make it ignore this.